requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
python-docx>=1.1.2
openpyxl>=3.1.2
